requirements.txt
Django==5.0.8
pillow==10.3.0
setuptools==70.1.1
django-debug-toolbar==4.4.6
django-ckeditor==6.7.0
djangorestframework==3.15.2
markdown==3.7
django-filter==24.3
drf-yasg[validation]
# temp
djangorestframework-simplejwt==5.3.1
